fix: address critical and high severity issues from PR review
Security fixes:
- Reject audit challenges where challenged_peer_id != self (oracle attack)
- Validate fetch response key matches requested key (data injection)
- Reject audit challenges exceeding MAX_AUDIT_CHALLENGE_KEYS (DoS)

Correctness fixes:
- Preserve bootstrap_claims across neighbor sync cycle transitions
- Shutdown replication engine before P2P node on shutdown
- Add pending_verify to enqueue_fetch dedup guard (Rule 8)
- Make FetchCandidate PartialEq/Ord consistent (BinaryHeap invariant)
- Change paid_list_check_indices from u16 to u32 (silent truncation)

Performance fixes:
- Make all_keys()/get_raw() async with spawn_blocking (tokio blocking)
- Release queues write lock before paid-list inserts in verification

+++ src/replication/protocol.rs
@@ -214,7 +214,7 @@ pub struct VerificationRequest {
     pub keys: Vec<XorName>,
     /// Which keys need paid-list status in addition to presence.
     /// Each value is an index into the `keys` vector.
-    pub paid_list_check_indices: Vec<u16>,
+    pub paid_list_check_indices: Vec<u32>,
 }
 
 /// Per-key verification result from a peer.
